SLA-1746 Remove old DropDown component
