--- src/core/ui/components/drop-down/content.tsx
@@ -1,7 +1,0 @@
-import styles from './styles.module.css';
-
-export const Content = ({ children }) => (
-  <div className={styles.content} data-testid={'Content'}>
-    {children}
-  </div>
-);
