--- src/core/ui/components/drop-down/primaryID.tsx
@@ -1,9 +1,0 @@
-export const PrimaryID = ({ children }) => (
-  <div style={primaryIdStyle}>{children}</div>
-);
-export const primaryIdStyle: React.CSSProperties = {
-  fontWeight: 500,
-  fontSize: '16px',
-  lineHeight: '20px',
-  textAlign: 'left',
-};
