+++ src/index.ts
@@ -8,7 +8,6 @@ import { supportedFontFamilies } from './core/ui/fonts';
 import { Account } from './shared/global';
 import { Buffer } from 'buffer';
 import { UserAccountSettings } from './core/ui/components/user_account_settings';
-import { DropDown } from './core/ui/components/drop-down/v1';
 
 if (typeof window !== 'undefined') {
   window.Buffer = Buffer;
@@ -27,7 +26,6 @@ export {
 export * from './core/ui';
 
 export {
-  DropDown,
   UserAccountSettings,
   Account,
   SlashAuth,
